+++ packages/calcite-components/src/components/flow-item/flow-item.stories.ts
@@ -1,19 +1,16 @@
 import { boolean, modesDarkDefault } from "../../../.storybook/utils";
 import { html } from "../../../support/formatting";
 import { ATTRIBUTES } from "../../../.storybook/resources";
+import { FlowItem } from "./flow-item";
 import { SLOTS } from "./resources";
 const { collapseDirection, scale } = ATTRIBUTES;
 
-interface FlowItemArgs {
-  closed: boolean;
-  disabled: boolean;
-  closable: boolean;
-  collapsible: boolean;
-  collapsed: boolean;
-  collapseDirection: string;
+interface FlowItemStoryArgs
+  extends Pick<
+    FlowItem,
+    "closed" | "disabled" | "closable" | "collapsible" | "collapsed" | "collapseDirection" | "loading" | "selected"
+  > {
   heightScale: string;
-  loading: boolean;
-  selected: boolean;
 }
 
 export default {
@@ -27,6 +24,7 @@ export default {
     collapseDirection: collapseDirection.defaultValue,
     heightScale: scale.defaultValue,
     loading: false,
+    selected: false,
   },
   argTypes: {
     collapseDirection: {
@@ -78,7 +76,7 @@ const flowItemContent = `${headerHTML}
   ${contentHTML}
   ${footerHTML}`;
 
-export const simple = (args: FlowItemArgs): string => html`
+export const simple = (args: FlowItemStoryArgs): string => html`
   <calcite-flow-item
     ${boolean("closed", args.closed)}
     ${boolean("disabled", args.disabled)}