+++ packages/calcite-components/src/components/pagination/pagination.stories.ts
@@ -2,15 +2,12 @@ import { locales, numberingSystems, defaultLocale, defaultNumberingSystem } from
 import { createBreakpointStories, modesDarkDefault } from "../../../.storybook/utils";
 import { html } from "../../../support/formatting";
 import { ATTRIBUTES } from "../../../.storybook/resources";
+import { Pagination } from "./pagination";
 const { scale } = ATTRIBUTES;
 
-interface PaginationArgs {
-  scale: string;
-  startItem: number;
+interface PaginationStoryArgs
+  extends Pick<Pagination, "scale" | "startItem" | "numberingSystem" | "totalItems" | "pageSize"> {
   lang: string;
-  numberingSystem: string;
-  totalItems: number;
-  pageSize: number;
 }
 
 export default {
@@ -44,7 +41,7 @@ export default {
   },
 };
 
-export const simple = (args: PaginationArgs): string => html`
+export const simple = (args: PaginationStoryArgs): string => html`
   <style>
     .sb-show-main.sb-main-centered #storybook-root {
       padding: 0 !important;