+++ packages/calcite-components/src/utils/floating-ui.ts
@@ -91,7 +91,7 @@ export const positionFloatingUI =
       overlayPositioning: Strategy;
       placement: LogicalPlacement;
       flipDisabled?: boolean;
-      flipPlacements?: EffectivePlacement[];
+      flipPlacements?: FlipPlacement[];
       offsetDistance?: number;
       offsetSkidding?: number;
       arrowEl?: SVGElement;
@@ -117,7 +117,7 @@ export const positionFloatingUI =
       middleware: getMiddleware({
         placement,
         flipDisabled,
-        flipPlacements,
+        flipPlacements: flipPlacements?.map((placement) => getEffectivePlacement(floatingEl, placement)),
         offsetDistance,
         offsetSkidding,
         arrowEl,
@@ -244,7 +244,9 @@ export const menuEffectivePlacements: EffectivePlacement[] = [
   "bottom-end",
 ];
 
-export const flipPlacements: EffectivePlacement[] = [
+export type FlipPlacement = Exclude<LogicalPlacement, "auto" | "auto-start" | "auto-end">;
+
+export const flipPlacements: FlipPlacement[] = [
   "top",
   "bottom",
   "right",
@@ -257,6 +259,12 @@ export const flipPlacements: EffectivePlacement[] = [
   "right-end",
   "left-start",
   "left-end",
+  "leading",
+  "trailing",
+  "leading-start",
+  "leading-end",
+  "trailing-start",
+  "trailing-end",
 ];
 
 export type MenuPlacement = Extract<
@@ -374,14 +382,14 @@ function getMiddleware({
   return [];
 }
 
-export function filterComputedPlacements(placements: string[], el: HTMLElement): EffectivePlacement[] {
+export function filterValidFlipPlacements(placements: string[], el: HTMLElement): EffectivePlacement[] {
   const filteredPlacements = placements.filter((placement: EffectivePlacement) =>
-    effectivePlacements.includes(placement),
+    flipPlacements.includes(placement),
   ) as EffectivePlacement[];
 
   if (filteredPlacements.length !== placements.length) {
     console.warn(
-      `${el.tagName}: Invalid value found in: flipPlacements. Try any of these: ${effectivePlacements
+      `${el.tagName}: Invalid value found in: flipPlacements. Try any of these: ${flipPlacements
         .map((placement) => `"${placement}"`)
         .join(", ")
         .trim()}`,